Blocker is now C++-free
Added `catch_pingus()` in PinguAction, so I don't have to port iteration
code to Céu yet. Also added `CatchedPingu` struct, pointer to which is
passed along with CATCH_PINGU event, so actions can implement catching.

There is only Blocker action which does catching, and as soon as
PinguHolder is ported to Céu, it would be replaced with interation over
a pool of Pingu objects.